user/requirements.txt
Flask==0.12
mock==2.0.0
ipdb==0.10.1
pytest==3.0.6
pika==0.10.0
psycopg2==2.6.1 
Flask-SQLAlchemy===2.1
Flask-Migrate==1.8.0
flask-marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.1
requests==2.18.1